% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assurance-package.R
\docType{package}
\name{assurance-package}
\alias{assurance-package}
\title{Compute assurance}
\description{
Perform assurance computations for a clinical trial, comparing means,
proportions, or hazards.
}
\details{
Estimate the unconditional probability of a study achieving its desired goal
by averaging the power of the study over the prior distribution of the
treatment effect. This is termed `assurance' by O'Hagan et al (2005), but
also see Spiegelhalter et al (200X) and the references therein.

The information provided to the function is used to construct a prior
density. For Gaussian data, the marginal density of the variance is taken to
be scaled inverse \eqn{\chi^2}{chi^2} and the conditional density of the
mean is taken to be Gaussian (see pages 74 - 75 of Gelman et al).

For binomial data, the marginal proportions are taken to be beta
distributions resulting from using Jeffreys' prior. That is, \eqn{p_1 \sim
B(x_1 + 0.5, m_1 - x_1 + 0.5)}{p1 ~ Beta(x1 + 0.5, m1 - x1 + 0.5} and
similarly for \eqn{p_2}{p2}.

For time-to-event data, the log of the hazard ratio is taken to be Gaussian
with known variance \eqn{1/x_1 + 1/x_2}{1/x1 + 1/x2} (see, for example,
Carroll, 2003).  Treating the variance as fixed fails to account for some
uncertainty, but in practice the loss is small because it is usual for the
number of events to be prespecified as part of the study design.
}
\references{
A. O'Hagan, J. W. Stevens and M. J. Campbell, Assurance in
clinical trial design, Pharmaceutical Statistics,4, 187 - 201, 2005

D. J. Spiegelhalter, K. R. Abrams and J. P. Myles, Bayesian Approaches to
Clinical Trials and Health-care Evaluation, Wiley, 2003

A. Gelman, J. B. Carlin, H. S. Stern and D. B. Rubin, Bayesian Data Analysis
(Second Edition), Chapman & Hall/CRC, 2004

K. J. Carroll, On the use and utility of the Weibull model in the analysis
of survival data, Controlled Clinical Trials, 24, 682 - 701, 2003
}
\author{
Harry Southworth, Paul Metcalfe

Maintainer: Paul Metcalfe <paul.metcalfe@astrazeneca.com>
}
\keyword{design}
