% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assurance.boot.R
\name{assurance.boot}
\alias{assurance.boot}
\title{Function for bootstrapping assurance values}
\usage{
assurance.boot(data, statistic = surv.statistic(total.events = 500),
  R = 500)
}
\arguments{
\item{data}{Data frame containing the data to be
bootstrapped. Will contain different columns depending on which
type of analysis we wish to analyse.}

\item{statistic}{Function that will perform the assurance value
calculation}

\item{R}{number of bootstrap replications (Default: 500)}
}
\description{
Function for bootstrapping assurance values
}

