% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assurance-package.R
\name{assurance.multiple.endpoints}
\alias{assurance.multiple.endpoints}
\title{Compute Assurance for Multiple Correlated Endpoints}
\arguments{
\item{earlyStudy}{Data defining the initial study.}

\item{laterStudy}{Information about the later study.}

\item{nsim}{Number of Monte-Carlo simulations (maybe unnecessary).}

\item{posteriorSample}{A sample from our posterior distribution.}

\item{laterSample}{Simulations of the later trial.}
}
\value{
\code{assurance.multiple.endpoints} returns a list of items:
NoAdjustment.Each returns (uncorrected) assurance values for each endpoint
separately. NoAdjustment.AtLeastOne returns the (uncorrected) assurance that
at least one endpoint will be statistically significant. NoAdjustmentAll
returns the (uncorrected) assurance that all endpoints will be statistically
significant. Currently no multiplicity adjustments are implemented.

\code{samplePosterior} returns an opaque object that contains samples from
the posterior.

\code{sampleLater} returns an opaque object that contains simulated results
for the later study.

\code{testLater} returns probability of success for the simulated later
study.

\code{sampleAndTestLater} returns probability of success for simulated later
study.
}
\description{
These generic functions provide the framework for assurance computations for
trials with multiple endpoints.
}
\details{
These generic functions tie together the whole assurance calculation for
multiple endpoints.  By default they do Monte-Carlo simulation using
\code{samplePosterior} (which samples from the posterior distribution),
\code{sampleLater} (which, given a sample from the posterior, simulates the
later study), and \code{sampleAndTestLater}.  \code{sampleAndTestLater} is
just the composition of \code{samplePosterior} and \code{testLater}.

\code{sampleAndTestLater} exists for the purposes of shortcut calculations
in which, given a sample from the posterior, the probability of later stage
success can be directly determined.
}
\author{
Paul Metcalfe, Mary Jenner, Daniel Dalevi
}
