% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assurance-package.R
\name{assurance}
\alias{assurance}
\alias{sampleAndTestLater}
\alias{sampleLater}
\alias{samplePosterior}
\alias{testLater}
\title{Compute Assurance}
\arguments{
\item{earlyStudy}{Data defining the initial study.}

\item{laterStudy}{Information about the later study.}

\item{nsim}{Number of Monte-Carlo simulations (maybe unnecessary).}

\item{posteriorSample}{A sample from our posterior distribution.}

\item{laterSample}{Simulations of the later trial.}
}
\value{
\code{assurance} returns a scalar; the calculated assurance.

\code{samplePosterior} returns an opaque object that contains samples from
the posterior.

\code{sampleLater} returns an opaque object that contains simulated results
for the later study.

\code{testLater} returns probability of success for the simulated later
study.

\code{sampleAndTestLater} returns probability of success for simulated later
study.
}
\description{
These generic functions provide the framework for assurance computations.
}
\details{
These generic functions tie together the whole assurance calculation.  By
default they do Monte-Carlo simulation using \code{samplePosterior} (which
samples from the posterior distribution), \code{sampleLater} (which, given a
sample from the posterior, simulates the later study), and
\code{sampleAndTestLater}.  \code{sampleAndTestLater} is just the
composition of \code{samplePosterior} and \code{testLater}.

\code{sampleAndTestLater} exists for the purposes of shortcut calculations
in which, given a sample from the posterior, the probability of later stage
success can be directly determined.
}
\examples{
count.data <-
  new.overdispersedPoisson(grp1.count=27,
                           grp1.overdispersion=1.6,
                           grp1.time.at.risk=97,
                           grp2.count=56,
                           grp2.overdispersion=1.6,
                           grp2.time.at.risk=82)

later.study <-
  new.overdispersedPoissonStudy(duration=56./52,
                                grp1.size=228,
                                grp2.size=228,
                                grp1.overdispersion=1.6,
                                grp2.overdispersion=1.6,
                                significance=0.04)

assurance(count.data, later.study, nsim=100000)
}
\author{
Paul Metcalfe
}

