% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assuranceNormal.R
\name{assuranceNormal}
\alias{assuranceNormal}
\title{Estimate the assurance for a two-arm trial with normally distributed data}
\usage{
assuranceNormal(
  fitDelta,
  fitPrecisionTmt,
  precisionCtrl = "iid",
  fitPrecisionCtrl = NULL,
  sigmaSqCtrl = NULL,
  pDeltaZero = 0,
  nControl = 25,
  nTreatment = nControl,
  deltaDist = "normal",
  precisionTmtDist = "gamma",
  precisionCtrlDist = precisionTmtDist,
  nRep = 10000,
  size = 0.05
)
}
\arguments{
\item{fitDelta}{an object of class \code{elicitation}
(obtained with \code{SHELF::fitdist()}), which
describes the elicited prior distribution for the treatment effect
\eqn{\delta}, conditional on \eqn{\delta \neq 0}.}

\item{fitPrecisionTmt}{an object of class \code{elicitation}, which
describes the elicited prior distribution for the treatment group precision
\eqn{\sigma_t^{-2}}.}

\item{precisionCtrl}{character string: one of \code{"iid"} for same distribution
as treatment group precision; \code{"separate"} if a different distribution is to be
used; \code{"fixed"} if the control group variance is to be fixed at a constant.}

\item{fitPrecisionCtrl}{an object of class \code{elicitation}, which
describes the elicited prior distribution for the control group precision
\eqn{\sigma_c^{-2}}. Only used if \code{precisionCtrl} is set to \code{separate}.}

\item{sigmaSqCtrl}{a fixed value to be used for the control group variance.
Only used if \code{precisionCtrl} is set to \code{"fixed"}.}

\item{pDeltaZero}{the prior probability that the treatment effect
\eqn{\delta = 0}.}

\item{nControl}{the sample size for the control arm.}

\item{nTreatment}{the sample size for the treatment arm.}

\item{deltaDist}{the chosen distribution for the treatment effect \eqn{\delta},
conditional on \eqn{\delta \neq 0}. One of \code{"normal"},
\code{"t"}, \code{"gamma"}, \code{"lognormal"}, \code{"logt"},
\code{"beta"}.}

\item{precisionTmtDist}{the chosen distribution for the treatment group precision.
Can be either \code{"gamma"} or \code{"lognormal"}}

\item{precisionCtrlDist}{the chosen distribution for the control group precision.
Can be either \code{"gamma"} or \code{"lognormal"}.
Ignored if \code{precisionCtrl} is set to \code{"fixed"}.}

\item{nRep}{the number of simulated trials used in the Monte Carlo estimate
of the assurance.}

\item{size}{the significance level used in the hypothesis test of no treatment
effect.}
}
\description{
Given elicited distributions for the treatment effect
and observation variance, this function will simulate observations
for a number of trials, and return the proportion of trials that result
in a statistically significant difference between the treatment and
control groups
}
