% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assuranceApp.R
\name{assuranceNormalApp}
\alias{assuranceNormalApp}
\title{Launch app for elicitation and assurance for normally distributed data}
\usage{
assuranceNormalApp()
}
\description{
Launch app for elicitation and assurance for normally distributed data
}
