% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assurance.boot.R
\name{binomial.statistic}
\alias{binomial.statistic}
\title{Calculates assurace for binomial data (patient having a reponse or
not) comparing two arms}
\usage{
binomial.statistic(n, alpha = 0.05, repl = 1e+05, C_ARM = "arm",
  C_EFFECT = "response", hurdle = NULL)
}
\arguments{
\item{n}{Total number of subjects in new study}

\item{alpha}{Significance level (default 0.05)}

\item{repl}{Number of simulations used in assurance calculation}

\item{C_ARM}{Column name of the arm data which is represented by 0 or 1 for active or control}

\item{C_EFFECT}{Column name for the effects (endpoint values of interest)}
}
\value{
A function that calculates the assurance values for gaussian data
}
\description{
Calculates assurace for binomial data (patient having a reponse or
not) comparing two arms
}

