% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assurance.boot.R
\name{check.arm}
\alias{check.arm}
\title{Function that will check the ARM data column for integrity}
\usage{
check.arm(data, C_ARM)
}
\description{
Function that will check the ARM data column for integrity
}

