% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{fit.nb}
\alias{fit.nb}
\title{Fit a negative binomial regression model to two-arm count data.}
\usage{
fit.nb(comparatorEvents, testEvents, duration)
}
\arguments{
\item{comparatorEvents}{count of events for each patient on comparator}

\item{testEvents}{count of events for each patient on active}

\item{duration}{duration of trial}
}
\description{
Fit a negative binomial regression model to two-arm count data.
}
