% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assurance.boot.R
\name{gauss.statistic}
\alias{gauss.statistic}
\title{Calculates assurace for gaussian data (difference in means)}
\usage{
gauss.statistic(n0, n1, alpha = 0.05, repl = 1e+05, C_ARM = "arm",
  C_EFFECT = "effect")
}
\arguments{
\item{n0}{Subjects in control arm in planned study}

\item{n1}{Subjects in active arm in planned study}

\item{alpha}{Significance level (default 0.05)}

\item{repl}{Number of simulations used in assurance calculation}

\item{C_ARM}{Column name of the arm data which is represented by 0 or 1 for active or control}

\item{C_EFFECT}{Column name for the effects (endpoint values of interest)}
}
\value{
A function that calculates the assurance values for gaussian data
}
\description{
Calculates assurace for gaussian data (difference in means)
}

