% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gaussian}
\alias{gaussian}
\alias{length,gaussianCommonSDParameters-method}
\alias{length,gaussianDifferentSDParameters-method}
\alias{samplePosterior,gaussianCommonSDInitialData,numeric-method}
\alias{samplePosterior,gaussianDifferenceSDInitialData,numeric-method}
\alias{treatmentEffect,gaussianCommonSDParameters-method}
\alias{treatmentEffect,gaussianDifferentSDParameters-method}
\alias{sampleLater,gaussianCommonSDParameters,twoArm-method}
\alias{sampleLater,gaussianDifferentSDParameters,twoArm-method}
\alias{testLater,gaussianCommonSDLater-method}
\alias{testLater,gaussianDifferentSDLater-method}
\title{Assurance for Gaussian endpoints}
\description{
Assurance can be calculated for Gaussian endpoints, either
assuming common std. dev. between arms or different
std. dev. between arms.
}
\details{
The prior distributions used are the improper
translation-invariant prior for treatment effect and the improper
scale invariant prior for the nuisance parameters of standard
deviation.
}
\examples{
initial.data <- new.gaussian(delta.mu=6,
                             sd1=10, sd2=20, m1=50, m2=20)
later.study <- new.twoArm(size=study.size(grp1.size=100,
                         grp2.size=200),
                      significance=0.05)
assurance(initial.data, later.study, 100000)
}
