% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assurance.boot.R
\name{kaplan.meier.plot}
\alias{kaplan.meier.plot}
\title{Plotting Kaplan Meier plot for survival data, used in examples.}
\usage{
kaplan.meier.plot(data, C_ARM = "TRTNUM", C_TIME = "D_DAYS",
  C_EVENT = "DEATH")
}
\description{
Plotting Kaplan Meier plot for survival data, used in examples.
}

