% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{new.binomialStudy}
\alias{new.binomialStudy}
\title{make a binomial study}
\usage{
new.binomialStudy(
  size,
  endpoint = c("cure", "mortality"),
  method = c("normal", "chisq", "wald", "bl", "ac"),
  significance = as.numeric(NA),
  hurdle = as.numeric(NA),
  margin = as.numeric(NA)
)
}
\description{
make a binomial study
}
