% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleInterim.R
\name{sampleInterim}
\alias{sampleInterim}
\title{Simulate posterior probabilities that the treatment is effective, given a new study}
\usage{
sampleInterim(
  fitDelta,
  fitPrecisionTmt,
  precisionCtrl = "iid",
  fitPrecisionCtrl = NULL,
  sigmaSqCtrl = NULL,
  pDeltaZero = 0,
  nControl = 25,
  nTreatment = nControl,
  deltaDist = "normal",
  precisionTmtDist = "gamma",
  precisionCtrlDist = precisionTmtDist,
  deltaThreshold = 0,
  diagnostics = FALSE,
  nTrials = 100,
  nIter = 1000,
  nAdapt = 1000,
  nBurn = 0,
  no_cores = parallel::detectCores() - 1
)
}
\arguments{
\item{fitDelta}{an object of class \code{elicitation}
(obtained with \code{SHELF::fitdist()}), which
describes the elicited prior distribution for the treatment effect
\eqn{\delta}, conditional on \eqn{\delta \neq 0}.}

\item{fitPrecisionTmt}{an object of class \code{elicitation}, which
describes the elicited prior distribution for the treatment group precision
\eqn{\sigma_t^{-2}}.}

\item{precisionCtrl}{character string: one of \code{"iid"} for same distribution
as treatment group precision; \code{"separate"} if a different distribution is to be
used; \code{"fixed"} if the control group variance is to be fixed at a constant.}

\item{fitPrecisionCtrl}{an object of class \code{elicitation}, which
describes the elicited prior distribution for the control group precision
\eqn{\sigma_c^{-2}}. Only used if \code{precisionCtrl} is set to \code{separate}.}

\item{sigmaSqCtrl}{a fixed value to be used for the control group variance.
Only used if \code{precisionCtrl} is set to \code{"fixed"}.}

\item{pDeltaZero}{the prior probability that the treatment effect
\eqn{\delta = 0}.}

\item{nControl}{the sample size for the control arm.}

\item{nTreatment}{the sample size for the treatment arm.}

\item{deltaDist}{the chosen distribution for the treatment effect \eqn{\delta},
conditional on \eqn{\delta \neq 0}. One of \code{"normal"},
\code{"t"}, \code{"gamma"}, \code{"lognormal"}, \code{"logt"},
\code{"beta"}.}

\item{precisionTmtDist}{the chosen distribution for the treatment group precision.
Can be either \code{"gamma"} or \code{"lognormal"}}

\item{precisionCtrlDist}{the chosen distribution for the control group precision.
Can be either \code{"gamma"} or \code{"lognormal"}.
Ignored if \code{precisionCtrl} is set to \code{"fixed"}.}

\item{deltaThreshold}{the minimum required value for the treatment effect: for each
trial the posterior probability \eqn{P(\delta >}\code{deltaThreshold}|data) will
be estimated.}

\item{diagnostics}{logical. Set to \code{TRUE} to produce convergence diagnostics
for one example chain.}

\item{nTrials}{number of trials to be simulated.}

\item{nIter}{number of MCMC iterations per simulated trial}

\item{nAdapt}{number of iterations per simulated trial used prior to choose the
proposal distributions. See \code{?rjags::adapt} for details}

\item{nBurn}{number of iterations discarded from each Markov chain as burn-in.}

\item{no_cores}{number of cores used in \code{parallel::parLapply}}
}
\value{
a vector of simulated posterior probabilities
\eqn{P(\delta >}\code{deltaThreshold}|data)
}
\description{
Simulates a number of trials, and for each simulated trial dataset, estimates
the posterior probability that the treatment is effective, given the simulated
data. This can be used to assess how informative a new trial would be, by
investigating how close the simulated posterior probabilities are to 0 or 1.
}
