% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assurance.boot.R
\name{simulate.binomial.trial}
\alias{simulate.binomial.trial}
\title{Simulate a gaussian trial, used in examples.}
\usage{
\method{simulate}{binomial.trial}(n0, n1, p0, p1, max0 = 1, max1 = 1)
}
\arguments{
\item{n0}{number of subjects in arm 0}

\item{n1}{number of subjects in arm 1}

\item{p0}{event probability in control}

\item{p1}{event probability in active}
}
\description{
Simulate a gaussian trial, used in examples.
}

