% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assurance.boot.R
\name{simulate.gauss.trial}
\alias{simulate.gauss.trial}
\title{Function that will simulate a gaussian trial, used in examples.}
\usage{
\method{simulate}{gauss.trial}(n0, n1, my0, my1, sd0, sd1)
}
\arguments{
\item{n0}{number of subjects in arm 0}

\item{n1}{number of subjects in arm 1}

\item{my0}{mean effect in control}

\item{my1}{mean effect in active}

\item{sd0}{sd in control}

\item{sd1}{sd in active}
}
\description{
Function that will simulate a gaussian trial, used in examples.
}

