% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single.arm.R
\name{single-arm}
\alias{assurance,singleArmInitialData,oneArmResponseRate,missing-method}
\alias{length,singleArmProbabilities-method}
\alias{sampleAndTestLater,singleArmProbabilities,oneArmResponseRate-method}
\alias{sampleLater,singleArmProbabilities,oneArmResponseRate-method}
\alias{samplePosterior,singleArmInitialData,numeric-method}
\alias{single-arm}
\alias{testLater,singleArmLater-method}
\alias{treatmentEffect,singleArmProbabilities-method}
\title{Assurance for single arm response rate}
\description{
Assurance can be calculated for response rate
}
\details{
Arbitrary beta distributions can be used as prior.
}
\examples{
initial.data <- new.singleArm(15, 30)
later <- new.oneArmResponseRate(100, 0.43)
# do it numerically
assurance(initial.data, later, 1000000)
# do it analytically
assurance(initial.data, later)
}

