% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assurance.boot.R
\name{surv.statistic}
\alias{surv.statistic}
\title{Calculates assurace for survival data, assumes time-event data is a function of arm (either 0 or 1)}
\usage{
surv.statistic(total.events, alpha = 0.05, repl = 1e+05, C_ARM = "TRTNUM",
  C_TIME = "D_DAYS", C_EVENT = "DEATH")
}
\arguments{
\item{total.events}{Number of events we est. will happen in later study}

\item{alpha}{Significance level (default 0.05)}

\item{repl}{Number of simulations used in assurance calculation}

\item{C_ARM}{Column name of the arm data which is represented by 0 or 1 for active or control}

\item{C_TIME}{Column name of the event times}

\item{C_EVENT}{Column name of the indicator for the event}
}
\value{
A function that calculates the assurance values for survival data
}
\description{
Calculates assurace for survival data, assumes time-event data is a function of arm (either 0 or 1)
}

