% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assurance-package.R
\name{bin}
\alias{bin}
\alias{new.binomial}
\alias{samplePosterior,binomialInitialData,numeric-method}
\alias{sampleLater,binomialProbabilities,twoArm-method}
\alias{testLater,binomialLater-method}
\alias{length,binomialProbabilities-method}
\title{Binomial assurance tests}
\arguments{
\item{x1}{The number of `successes' observed in treatment group 1 in the
existing study.}

\item{m1}{The number of observations from which \code{x1} successes were
observed}

\item{x2}{The number of `successes' observed in treatment group 2 in the
existing study.}

\item{m2}{The number of observations from which \code{x2} successes were
observed}
}
\description{
Creates an object representing results of initial study.
}
\examples{

bin.data <- new.binomial(x1=18, m1=20, x2=15, m2=20)
}
\author{
Paul Metcalfe
}
