% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assurance-package.R, R/binomial.R, R/internal.R
\name{internal}
\alias{.ni.ac}
\alias{.ni.bl}
\alias{.ni.test}
\alias{.ni.wald}
\alias{assert}
\alias{assurance-internal.Rd}
\alias{check.significance}
\alias{gaussianCommonSDParameters}
\alias{gaussianCommonSDParameters-class}
\alias{gaussianDifferentSDParameters}
\alias{gaussianDifferentSDParameters-class}
\alias{internal}
\alias{is.natural}
\alias{is.positive}
\alias{is.positive.scalar}
\alias{is.probability}
\alias{is.wholenumber}
\alias{length}
\alias{length-method}
\alias{testLater}
\title{Internal classes and methods for assurance package}
\usage{
.ni.test(effect, se, alpha, margin, direction)

.ni.wald(event.1, size.1, event.2, size.2, ...)

.ni.ac(event.1, size.1, event.2, size.2, ...)

.ni.bl(event.1, size.1, event.2, size.2, ...)

assert(condition, msg = "assertion failed")

is.wholenumber(num)

is.positive(num)

is.positive.scalar(x)

is.natural(num)

is.probability(value)

check.significance(alpha, expr, frame = parent.frame())
}
\description{
Internal assurance functions.
}
\details{
These internal \pkg{assurance} functions are not usually called directly by
the user.
}
\keyword{internal}

